Added if statement for first.width
This change will solve a bug that will happen when there is no `first.width` available. 

The code will now add one image in the `src-set` with `1w,` which does not work in Safari on the iPad and with this change it will not add any `src-set` if there are no images available to put in there.